backend/requirements.txt
# Use CPU-only PyTorch (avoids downloading 5GB+ GPU version)
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision

# Flask and extensions
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
flask-migrate==4.0.5
flask-sqlalchemy==3.1.1

# Database
psycopg2-binary==2.9.9  # Uncomment for PostgreSQL
sqlalchemy>=2.0.35

# AI/ML
groq>=0.11.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0
sentence-transformers>=2.7.0
faiss-cpu>=1.9.0
chromadb>=0.4.22

# PDF Processing
pypdf>=4.0.0

# Utilities
python-dotenv==1.0.0
numpy>=1.26.2
tiktoken>=0.7.0
gunicorn==21.2.0
redis>=5.0.1
